Ensure ia32 simulation actually runs on i386
This ensures that when we simulat ia32 builds we simulate them on the
qemu i386 emulator rather than the x86_64 simulator.

This requires we explicitly disable the 'long mode' (64-bit) and
'syscall' options when not in use in the simulation script.

This most usefully ensures that when we attempt to connect the simulated
ia32 build to a debugger, that debugger actually uses correct word sizes
and can properly disassemble and list code.

Additionally, the i386 simulator does not support the MSR mechanism for
setting the FS or GS base, so we use the GDT instead.

This also means that ApplyCommonSimulationSettings must be passed the
KernelSel4Arch rather than just KernelArch such that it can distinguish
x86_64 and ia32.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>
GitOrigin-RevId: dcdd5653e0ad128ab0b112324ef3f33fc647b7c2