Add reset CNTVOFF for the arm generic timer
When the kernel runs in non-hyp mode (EL1) and uses the arm generic
timer, it uses the virtual counter which is the physical counter -
virtual offset. The virtual offset has no defined reset value, can only
be set from EL2 (i.e. in the bootloader), and must be reset separately
for each code.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>
GitOrigin-RevId: 9b9a705d8156b6ac186e2fe6747b6600c2eacc5b