Reset CNTVOFF on all arm generic timer platforms
When the kernel runs in EL1 on a platform with the arm generic timer the
virtual counter is used. MCS can only correctly schedule in the first
3/8ths of representable time so must ensure that the offset is reset for
the virtual timer while the bootloader is still in EL2 as EL1 cannot
change the offset used to generate the virtual counter value.

Signed-off-by: Curtis Millar <curtis.millar@data61.csiro.au>
GitOrigin-RevId: b64893955c68f5add9d14d1e788abbfbfae18252