cmake,cpio: Don't create folder at source location
When making a CPIO archive, the script tries to set the filetimes to 0
to allow reproducible builds. It needs to modify the input file to do
this and so creates a copy.  Now this copy is stored in a temp folder
that is relative to the destination path instead of input file source
path. This prevents race conditions when trying to create multiple CPIO
archives from the same source files, and also avoids complications
arising from trying to modify a different module's build directory after
it was built.

Signed-off-by: Kent McLeod <kent@kry10.com>
GitOrigin-RevId: 85ca1ead1c4666560e91fa52579b4db43ebec5c0