elfloader/shoehorn: cleanup sel4test fudge factor hack
elfloader uses shoehorn to find a memory location to load build
artifacts. There was a hack to artificially increase the required space
by 4MiB for sel4test. This breaks systems with limited memory so move
this hack to a shoehorn command line option and add cmake glue to enable
it to be set in the sel4test build glue.

Bug: 197745020

Change-Id: I146976ae23cd4fcdd7a422d9c1643428e3596efb
GitOrigin-RevId: 1a1c2c7e5bcbe9744afc1bbc3e25271af7c763c8